Moved common message box functions to a common file.
Common functions will be used in other tests.
