+++ tests/kernel/iop/messagebox/Makefile
@@ -1,6 +1,6 @@
 TARGETS = send
 
-EXTRA_OBJS = imports.o
+EXTRA_OBJS = imports.o mbx-common.o
 
 COMMON_DIR=../../../../common
 include $(COMMON_DIR)/common-iop.mk
